refactor(channels): preserve architecture budgets
