+++ src/lib/state/registry-messaging.ts
@@ -5,6 +5,7 @@ import { hydrateDerivedSandboxMessagingPlanFields } from "../messaging/hydration
 import type { SandboxMessagingPlan } from "../messaging/manifest";
 import { compactSandboxMessagingPlanForPersistence } from "../messaging/persistence";
 import {
+  getActiveChannelIdsFromPlan,
   getConfiguredChannelIdsFromPlan,
   getDisabledChannelIdsFromPlan,
   parseSandboxMessagingPlan,
@@ -66,6 +67,12 @@ export function getConfiguredMessagingChannelsFromEntry(
   entry: EntryWithMessaging | null | undefined,
 ): string[] {
   return getConfiguredChannelIdsFromPlan(getMessagingPlanFromEntry(entry));
+}
+
+export function getActiveMessagingChannelsFromEntry(
+  entry: EntryWithMessaging | null | undefined,
+): string[] {
+  return getActiveChannelIdsFromPlan(getMessagingPlanFromEntry(entry));
 }
 
 export function getDisabledMessagingChannelsFromEntry(
